requirements.txt
Django == 3.0.7
Pillow == 6.2.2
djangorestframework
django-cors-headers
